refactor: migrate domain types to BigDecimal for precision (#4)
- Replace primitive types with BigDecimal in Distance, Time, Pace, Speed
- Add BigDecimalBased interface for consistent value comparison
- Introduce UnitConversions utility with precise conversion constants
- Update all calculations to use BigDecimal arithmetic
- Add test utilities for BigDecimal equality comparison
- Ensure mathematical precision for running calculations

+++ app/src/main/java/me/fetsh/pacecalculator/ui/Screen.kt
@@ -35,6 +35,7 @@ import me.fetsh.pacecalculator.domain.Speed
 import me.fetsh.pacecalculator.domain.Time
 import me.fetsh.pacecalculator.domain.TimeParts
 import me.fetsh.pacecalculator.domain.TimePrecision
+import java.math.BigDecimal
 
 /**
  * Playground screen wiring domain inputs (Distance, Pace, Time, Speed) to the ViewModel.
@@ -213,7 +214,7 @@ fun Screen(viewModel: PaceCalculatorViewModel = viewModel()) {
                     .toString(),
             onValueChange = {
                 it.toDoubleOrNull()?.let { spd ->
-                    viewModel.processMessage(Msg.SpeedChanged(Speed(spd), uiState.speedUnit))
+                    viewModel.processMessage(Msg.SpeedChanged(Speed(BigDecimal.valueOf(spd)), uiState.speedUnit))
                 }
             },
             keyboardOptions = KeyboardOptions.Default.copy(keyboardType = KeyboardType.Number),
